requirements.txt
# Core dependencies for Multicamera Motion Capture System
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
opencv-python>=4.8.0
mediapipe>=0.10.0
numpy>=1.24.0
scipy>=1.11.0
onnxruntime-gpu>=1.16.0
matplotlib>=3.7.0
PyYAML>=6.0
websockets>=15.0.0
